Create assignment_2
1. Write a Java program to create a List containing a list of items of type
String and use for- --each loop to print the items of the list.
2. Write a Java program to create List containing list of items and use
ListIterator interface to print items present in the list. Also print the list in
reverse/ backward direction
3. Write a Java program to create a Set containing list of items of type
String and print the items in the list using Iterator interface. Also print the
set elements in reverse/ backward direction
4. Write a Java program using Set interface containing list of items and
perform the following operations:
a. Add items in the set.
b. Insert items of one set in to other set.
c. Remove items from the set
d. Search the specified item in the set
5. Write a Java program using Map interface containing list of items having
keys and associated values and perform the following operations:
A. Add items in the map.
B. Remove items from the map
C. Search specific key from the map
D. Get value of the specified key
E. Insert map elements of one map in to other map.
F. Print all keys and values of the map.
6. Write a Java program using Lambda Expression to print ”Hello World”.
7.Write a Java program using Lambda Expression with single parameters.
8.Write a Java program using Lambda Expression with multiple parameters to
add two numbers.
9. Write a Java program using Lambda Expression to calculate the following:
1. Convert Fahrenheit to Celcius
2. Convert Kilometers to Metres
10.Write a Java program using Lambda Expression with return keyword.Find
average of two numbers.
11.Write a Java program using Lambda Expression to concatenate two
strings.